pyproject.toml
[project]
name = "skin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "torch>=2.8.0",
    "ultralytics>=8.3.197",
    "uvicorn>=0.35.0",
]
